+++ src/TonemappingOperators/mai11/compression_tmo.h
@@ -39,8 +39,8 @@
 namespace mai {
 class CompressionTMO {
    public:
-    void tonemap(const float *R_in, const float *G_in, float *B_in, int width,
-                 int height, float *R_out, float *G_out, float *B_out,
+    void tonemap(const float *R_in, const float *G_in, float *B_in, size_t width,
+                 size_t height, float *R_out, float *G_out, float *B_out,
                  const float *L_in, pfs::Progress &ph);
 };
 }
